Move all existing tests to a 'unit' tests subpackage
