+++ Makefile
@@ -11,7 +11,10 @@ clean:
 	rm -rf build/ dist/ src/*.egg-info **/__pycache__ .coverage .pytest_cache/ .ruff_cache/
 
 test:
-	pytest
+	pytest tests/unit
+
+test-all:
+	pytest tests
 
 lint:
 	black --check src tests
