+++ tests/unit/models/test_models_classic_categories.py
