+++ tests/unit/models/test_models_classic_computer_groups.py
