+++ tests/unit/models/test_models_classic_network_segments.py
